dump color mapping to uart and bigger diagram

+++ software/gtia2dvi/gfx/gfx.h
@@ -26,7 +26,7 @@ uint8_t framebuf[FRAME_WIDTH * FRAME_HEIGHT];
 
 static void inline fill_screen(int8_t color)
 {
-    uint16_t *buf = (uint32_t *)framebuf;
+    uint16_t *buf = (uint16_t *)framebuf;
     uint16_t count = FRAME_WIDTH * FRAME_HEIGHT / 2;
     uint16_t val = (color << 16) + color;
     while (count--)
